Implemented Compose, Do_Nothing. Created imports.py. Added Tests.
Compose(A,B) creates a synchronous algorithm that runs A and B in parallel synchrony.
Do_Nothing is a synchronous algorithm that immediately terminates.
imports.py is just a grouping of relevant imports.
Additional tests for Do_Nothing and Compose.
Also implemented Tester.begin_testing(), to be called at the beginning of a test suite. Resets number of tests run to 0.